+++ src/SmartMap.php
@@ -41,6 +41,9 @@ use doublesecretagency\smartmap\listeners\GetAddressFieldSchema;
 class SmartMap extends Plugin
 {
 
+    /** @const string Root URL for documentation. */
+    const DOCS_URL = 'https://www.doublesecretagency.com/plugins/smart-map/docs';
+
     /** @event DetectLocationEvent The event that is triggered after the user's location has been detected. */
     const EVENT_AFTER_DETECT_LOCATION = 'afterDetectLocation';
 
@@ -142,7 +145,7 @@ class SmartMap extends Plugin
         return $view->renderTemplate('smart-map/settings', [
             'settings' => $this->getSettings(),
             'overrideKeys' => $overrideKeys,
-            'docsUrl' => $this->documentationUrl,
+            'docsUrl' => static::DOCS_URL,
         ]);
     }