+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- Fixed links to documentation.
+
 ## 3.2.1 - 2018-12-14
 
 ### Added